backend/requirements.txt
flask==2.3.3
flask-cors==6.0.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.7.2
werkzeug==2.3.7
reportlab==4.4.2 